refactored litererally everything
