+++ lib/converter.js
@@ -1,0 +1,50 @@
+#!/usr/bin/env node
+const fs = require("fs");
+const showdown = require("showdown");
+
+const card = require("./rules");
+
+// initialize shodown
+showdown.extension("anki", card.card);
+const converter = new showdown.Converter({
+  extensions: ["anki"]
+});
+
+// showdown options
+showdown.setOption("disableForced4SpacesIndentedSublists", true);
+converter.setOption("noHeaderId", true);
+
+// card preferences
+const cardLeft = new RegExp("<!--", "g");
+const cardRight = new RegExp("-->", "g");
+const delimiter = "~";
+
+// write data function
+function writeData(outputFile, inputData) {
+  // console.log(`Saving data to: ${outputFile}`);
+  inputData = inputData + "\n";
+  fs.appendFileSync(outputFile, inputData);
+  // console.log(`Finished writing to ${outputFile}`);
+}
+
+// converts data from MD to HTML
+function convertFile(input, output) {
+  let mdContent = fs.readFileSync(input, "utf-8");
+
+  htmlContent = converter.makeHtml(mdContent);
+
+  let firstCard = "<!-- ignore -->";
+  htmlContent = firstCard + htmlContent;
+  htmlContent = htmlContent.replace(/~/g, ""); // cleans delimiter
+  htmlContent = htmlContent.replace(/\n/g, ""); // replaces linebreak
+  htmlContent = htmlContent.replace(cardLeft, "\n"); // replaces cardleft
+  htmlContent = htmlContent.replace(cardRight, delimiter); // replaces cardright
+  htmlContent = htmlContent.replace(/<p>/s, ""); // replaces beginning paragraph tag
+  htmlContent = htmlContent.replace(/<\/p>/s, ""); // replaces end paragaph tag
+
+  writeData(output, htmlContent);
+}
+
+module.exports = {
+  convertFile
+};
