--- main/Pain and Temperature.md
@@ -1,254 +1,0 @@
-# Anterolateral System
-
-{{3 modalities of Anterolateral system. What is used to test ALS. What is spinal level}}
-
-![](Pain and Temperature/Slide2.JPG)
-
-- pain: not a needle, just a pin prick
-- temperature: not usually tested
-- crude touch: not discriminative. ==You just know something's== happening; also itch, tickle, etc.
-- spinal level: slang for lowest dermatome level where pinprick is still intact. E.g. sensory lvl of T6 has sensation intact down to T6 but not below
-
-{{1st order afferents, 2nd order neurons, 3 distinct tracts for ALS. How is it different from DCML}}
-
-![](Pain and Temperature/Slide2.JPG)
-
-# Spinothalamic tract
-
-## Modality
-
-{{ignore}}
-
-![](https://i.imgur.com/YGGrNEq.png)
-
-## Tract
-
-{{spinothalamic tr. overview, neurons, tracts, final destiny}}
-
-![](Pain and Temperature/Slide5.JPG)
-
-- Enter cord and ascend 1-2 segments right away in dorsal horn known as Lissauer's tract
-
-{{spinothalamic tr. synapse and enkephalinergic interneurons in spinal cord}}
-
-![](Pain and Temperature/Slide7.JPG)
-
-- C and A delta synapse in lamina 1, where enkephalin mediated inhibition occur
-- Know Substantia gelatinosa, where **enkephalingeic interneurons** are
-
-![](https://i.imgur.com/KOqOK4R.jpg)
-
-## Histology
-
-{{Spinothalamic tr. histology}}
-
-### Lower Cervical
-
-![](https://i.imgur.com/YiaWE5G.jpg)
-
-- most distal sacral, most medial for upper extremity
-- In lesions, sacral will be last dermatome to be affected (saddle)
-- central lesion at ant. white commissure can have bilateral results
-
-### Caudal Medulla
-
-![](https://i.imgur.com/lKufIsm.jpg)
-
-### Caudal Pons
-
-![](https://i.imgur.com/f2diuYm.jpg)
-
-### Pontomesencephalic Junction
-
-![](https://i.imgur.com/chA8gCB.jpg)
-
-### Midbrain
-
-![](https://i.imgur.com/AFCOf2t.jpg)
-
-![](https://i.imgur.com/D4ikB3a.jpg)
-
-### Thalamus
-
-![](https://i.imgur.com/8bA9Gqr.jpg)
-
-![](https://i.imgur.com/IXQPtry.jpg)
-
-- same as DCML
-
-![](https://i.imgur.com/nGm253R.jpg)
-
-# Spinoreticular pathway
-
-{{spinoreticular pathway mediates, overview, neurons and tracts}}
-
-![](https://i.imgur.com/XCdxKv7.jpg)
-
-- mediates emotional reaction to pain
-- spinoreticular, reticulothalamic, thalamocortical tracts
-
-## Histology
-
-{{spinoreticular tr. histology}}
-
-- spinoreticular
-
-![](https://i.imgur.com/5k4fszf.png)
-
-![](https://i.imgur.com/5E9B04S.png)
-
-
-
-![](https://i.imgur.com/Jh2zj3r.jpg)
-
-![](https://i.imgur.com/YJiIuPu.jpg)
-
-![](https://i.imgur.com/e0prgZV.jpg)
-
-{{thalamocortical projection function}}
-
-![](https://i.imgur.com/si3Ckw2.png)
-
-{{intralaminar nuclei thalamus location}}
-
-![](https://i.imgur.com/Ts5Y5OR.jpg)
-
-- white arrow; centralmedian example
-
-# Spinomesencephalic tr
-
-{{spinomesencephalic tr. ascending overview, function, neurons, tracts, fibers}}
-
-![](https://i.imgur.com/aRUPCvZ.jpg)
-
-- pain input activates co pathway (spinomesencephalic) that leads to descending neurons that limit pain
-
-## Histology
-
-{{spinomesencephalic tr. ascending pathway histology}}
-
-- Periaquaductal gray
-
-![](https://i.imgur.com/bpJaqyA.png)
-
-
-
-![](https://i.imgur.com/n5cWQ3J.jpg)
-
-- spinomesencephalic synapse on small PAG interneurons that produce peptides
-- activation of descending Glutaminergic neurons, descend to lower levels of brainstem
-
-
-## Descending pathway
-
-{{spinomesencephalic tr. descending pathway. What are 2 pathways. Where does the final inhibition happen}}
-
-![](https://i.imgur.com/ySHtgP6.jpg)
-
-- NE and seratonergic pathways
-
-![](https://i.imgur.com/i1lDyIP.jpg)
-
-- All happening in dorsal horn and substantia gelatinosa
-
-{{3 CNS sites where opioid peptides operate for pain}}
-
-![](https://i.imgur.com/zXN103V.png)
-
-## Descending NE Mechanism
-
-{{descending NE mechanism pathway, neuron, MOA, and result}}
-
-![](https://i.imgur.com/l1hLRQp.jpg)
-
-- activates the only NE neuron in the brain
-- axoaxonic synnapse and inhibit presynaptic neurons
-- result: gates K+ efflux = hyperpolarized presynapse = harder to depolarize
-- location of action for low dose tricyclic antidepressants
-
-## Descending Serotonergic Mechanism
-
-{{Descending serotonergic mechanism pathway, neuron, MOA, and result}}
-
-![](https://i.imgur.com/mbXmr7u.jpg)
-
-- Some SSRI can inhibit NE uptake and stimulate 5-HTR = synergistic effect, raise pain threshold
-
-{{substantia gelatinosa histology}}
-
-- substantia gelatinosa
-
-![](https://i.imgur.com/6ykcPYX.png)
-
-
-
-![](https://i.imgur.com/WScrqCC.jpg)
-
-- light area: high concentration of enkephalinergic neurons
-
-{{cannabinoids function and MOA}}
-
-![](https://i.imgur.com/GVrLVoS.jpg)
-
-{{limbic system and pain pathway}}
-
-![](https://i.imgur.com/v1ZjXMn.jpg)
-
-- People in fight or flight  don't feel pain until later
-- also present in kids before getting shots. Thus really complicated system
-
-{{locus ceruleus and raphe nuclei location}}
-
-![](https://i.imgur.com/10P9vjH.jpg)
-
-# Pain
-
-{{what is psychosomatic pain. 2 defective pathways. Why is it linked to depression. Example disease. Drug used to treat and MOA}}
-
-![](https://i.imgur.com/OGoH5jz.jpg)
-
-- mood disorder co present with lower pain threshold
-
-{{what is neuropathic pain. Examples. Surgical treatment}}
-
-![](https://i.imgur.com/QMZzLfI.jpg)
-
-- ignore MS, diabetic neuropathy, V/IX neuralgias
-
-# Placebo
-
-{{Placebo effect cause, pathway, result}}
-
-![](https://i.imgur.com/03CWlRe.jpg)
-
-- begins with primary senses, info relayed to reward center, inform PAG to elevate descending system
-
-# Nocebo
-
-{{Nocebo effect MOA}}
-
-![](https://i.imgur.com/mXM0s23.jpg)
-
-# Visceral Pain
-
-{{visceral pain overview pathway, neurons, tracts}}
-
-![](https://i.imgur.com/lO4SfG8.jpg)
-
-{{visceral pain histology}}
-
-![](https://i.imgur.com/6ykcPYX.png)
-
-![](https://i.imgur.com/YJjeZ4u.png)
-
-![](https://i.imgur.com/vKVWShy.jpg)
-
-- ALS does not control visceral pain. 
-- Visceral lumbar/abdominal 1 afferent synapse at lower level of dorsal horn and ascenda via PDCN in FG
-
-![](https://i.imgur.com/Plvi6oj.jpg)
-
-{{PSDCN down regulated by what drug, how. How are pain alleviated in patients with bowel cancer}}
-
-![](https://i.imgur.com/79DeNA5.jpg)
-
